requirements.txt
absl-py==0.10.0
alabaster==0.7.12
astroid==2.4.2
astunparse==1.6.3
atomicwrites==1.4.0
attrs==20.2.0
autopep8==1.5.4
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
bleach==3.1.5
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.3
cryptography==3.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
docutils==0.16
entrypoints==0.3
flake8==3.8.3
Flask==1.1.2
gast==0.3.3
google-auth==1.21.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
gunicorn==20.0.4
h5py==2.10.0
helpdev==0.7.1
idna==2.10
imagesize==1.2.0
importlib-metadata==1.7.0
intervaltree==3.1.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.5.1
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
Keras==2.4.3
Keras-Preprocessing==1.1.2
keyring==21.4.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
newrelic==5.18.0.148
numpy==1.18.5
numpydoc==1.1.0
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.1
pandocfilters==1.4.2
paramiko==2.7.2
parso==0.7.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.9.0
pluggy==0.13.1
prompt-toolkit==3.0.7
protobuf==3.13.0
psutil==5.7.2
ptyprocess==0.6.0
py4j==0.10.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.6.1
pylint==2.6.0
PyNaCl==1.4.0
pyodbc==4.0.30
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==12.8.1
PyQtWebEngine==5.12.1
pyrsistent==0.16.0
pyspark==2.4.0
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.34.1
pytz==2020.1
pywin32==228
pywin32-ctypes==0.2.0
PyYAML==5.3.1
pyzmq==19.0.2
QDarkStyle==2.8.1
QtAwesome==0.7.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rope==0.17.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.11.0
six==1.15.0
sklearn==0.0
snowballstemmer==2.0.0
sortedcontainers==2.2.2
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==4.1.5
spyder-kernels==1.9.4
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
tensorflow-gpu-estimator==2.3.0
termcolor==1.1.0
testpath==0.4.4
tflearn==0.3.2
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
traitlets==5.0.4
typed-ast==1.4.1
ujson==3.2.0
urllib3==1.25.10
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
yapf==0.30.0
zipp==3.1.0
